Have vim-addon-manager manage Vim plugins.
This includes the VAM auto-installer script in the .vimrc. When starting
(G)Vim in a new environment, it will attempt to fetch VAM using Git,
perform an auto-installation, and then proceed to install all the other
plugins using VAM.

I decided to explicitly not track VAM or plugins fetched by it in my Git
repository because of the fact that VAM likes to fetch plugins via VCSes
including Git and Bazaar, and I don't want to use Git submodules, nor
check in the contents of .bzr and .hg repositories fetched by VAM. As a
consequence, there will be a small startup cost when getting Vim going
in a different environment, but I think this is the correct solution.