Fixes "stack level too deep" issue.
I had an issue with LustyExplorer raising a "stack level too deep"
error.

https://github.com/sjbach/lusty/issues/67#issuecomment-4272186

It turns out it had to do with how I was calling the VAM setup function,
within an autocommand for "VimEnter". I'm not sure exactly what the
interaction was with LustyExplorer [Does this repeatedly activate the
VimEnter signal?] but changing it to just a straight call upon reading
the .vimrc works to fix the issue.